requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
packaging==26.0
pytz==2025.2
PyYAML==6.0.3
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
tzdata==2025.3
uritemplate==4.2.0
